String.indexOf: the type of the char argument is `int`. Whatever.

+++ modules/common/src/main/base/PimpedUtils.scala
@@ -43,10 +43,10 @@ final class PimpedOption[A](private val self: Option[A]) extends AnyVal {
 final class PimpedString(private val s: String) extends AnyVal {
 
   def replaceIf(t: Char, r: Char): String =
-    if (s.indexOf(t) >= 0) s.replace(t, r) else s
+    if (s.indexOf(t.toInt) >= 0) s.replace(t, r) else s
 
   def replaceIf(t: Char, r: CharSequence): String =
-    if (s.indexOf(t) >= 0) s.replace(String.valueOf(t), r) else s
+    if (s.indexOf(t.toInt) >= 0) s.replace(String.valueOf(t), r) else s
 
   def replaceIf(t: CharSequence, r: CharSequence): String =
     if (s.contains(t)) s.replace(t, r) else s
